Move IdentifierContext to kwasm.runtime
