+++ library/src/main/java/kwasm/format/text/Identifier.kt
@@ -15,7 +15,6 @@
 package kwasm.format.text
 
 import kwasm.ast.Identifier
-import kwasm.ast.IdentifierContext
 import kwasm.format.ParseContext
 import kwasm.format.ParseException
